+++ example/lib/settings.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_widgetz/flutter_widgetz.dart';
 
-// TODO: calling the callbacks is not working
 class SettingsView extends StatefulWidget {
   const SettingsView({
     super.key,
@@ -30,13 +29,13 @@ class _SettingsViewState extends State<SettingsView>
           title: const Text('Toggle Dark Mode'),
           leading: const Icon(Icons.dark_mode),
           value: false,
-          onChanged: (_) => widget.onDarkModeChanged,
+          onChanged: (_) => widget.onDarkModeChanged(),
         ),
         SettingsTile.switched(
           title: const Text('Show Material Grid'),
           leading: const Icon(Icons.grid_3x3),
           value: false,
-          onChanged: (_) => widget.onShowMaterialGrid,
+          onChanged: (_) => widget.onShowMaterialGrid(),
         ),
         SettingsTile.timeDilation(
           context: context,