+++ lib/src/auth_button.dart
@@ -144,7 +144,6 @@ class AuthButton extends StatelessWidget {
         padding: const EdgeInsets.only(right: 8.0),
       ),
       child: Container(
-        // TODO: A RenderFlex overflowed by 8.0 pixels on the right
         constraints: BoxConstraints(
           maxHeight: _height,
           minWidth: minWidth,